% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{sbfx_backup_pg}
\alias{sbfx_backup_pg}
\title{Save PostgreSQL backup}
\usage{
sbfx_backup_pg(
  db_dump_name = subfoldr2::sbf_get_db_name(),
  sub = subfoldr2::sbf_get_sub(),
  main = subfoldr2::sbf_get_main(),
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.config_value", "default")
)
}
\arguments{
\item{db_dump_name}{A string of the name for the database backup file.}

\item{sub}{A string specifying the path to the sub folder (by default the
current sub folder).}

\item{main}{A string specifying the path to the main folder (by default the
current main folder)}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option
and uses \code{NULL} if no value supplied.}

\item{config_value}{A string of the name of value. The default value grabs
the value from the psql.config_value option and uses \code{"default"} if no
value is supplied.}
}
\value{
TRUE (or errors).
}
\description{
Save a copy of your database in plain text format. This saves all SQL code to
recreate the structure and data.
}
\details{
Wrapper on \code{psql::psql_backup()}.
}
\examples{
\dontrun{
sbfx_backup_pg()
sbfx_backup_pg("database-22")
}
}
\seealso{
Other postgresql functions: 
\code{\link{sbfx_close_pg}()},
\code{\link{sbfx_create_pg}()},
\code{\link{sbfx_execute_pg}()},
\code{\link{sbfx_get_config_file}()},
\code{\link{sbfx_get_config_value}()},
\code{\link{sbfx_get_schema}()},
\code{\link{sbfx_list_tables_pg}()},
\code{\link{sbfx_load_data_from_pg}()},
\code{\link{sbfx_load_datas_from_pg}()},
\code{\link{sbfx_open_pg}()},
\code{\link{sbfx_reset_config_file}()},
\code{\link{sbfx_reset_config_value}()},
\code{\link{sbfx_reset_schema}()},
\code{\link{sbfx_save_data_to_pg}()},
\code{\link{sbfx_set_config_file}()},
\code{\link{sbfx_set_config_value}()},
\code{\link{sbfx_set_schema}()}
}
\concept{postgresql functions}
