% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{sbfx_create_pg}
\alias{sbfx_create_pg}
\title{Create PostgreSQL database}
\usage{
sbfx_create_pg(
  dbname,
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.config_value", "default")
)
}
\arguments{
\item{dbname}{A string of the name of the new database to create.}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option
and uses \code{NULL} if no value supplied.}

\item{config_value}{A string of the name of value. The default value grabs
the value from the psql.config_value option and uses \code{"default"} if no
value is supplied.}
}
\value{
TRUE (or errors).
}
\description{
Create a new PostgreSQL database.
}
\details{
Wrapper on \code{psql::psql_create_db()}.
}
\examples{
\dontrun{
sbfx_create_pg("new_database")
sbfx_create_pg("new_database", config_path = "keys/config.yml")
}
}
\seealso{
Other postgresql functions: 
\code{\link{sbfx_backup_pg}()},
\code{\link{sbfx_close_pg}()},
\code{\link{sbfx_execute_pg}()},
\code{\link{sbfx_get_config_file}()},
\code{\link{sbfx_get_config_value}()},
\code{\link{sbfx_get_schema}()},
\code{\link{sbfx_list_tables_pg}()},
\code{\link{sbfx_load_data_from_pg}()},
\code{\link{sbfx_load_datas_from_pg}()},
\code{\link{sbfx_open_pg}()},
\code{\link{sbfx_reset_config_file}()},
\code{\link{sbfx_reset_config_value}()},
\code{\link{sbfx_reset_schema}()},
\code{\link{sbfx_save_data_to_pg}()},
\code{\link{sbfx_set_config_file}()},
\code{\link{sbfx_set_config_value}()},
\code{\link{sbfx_set_schema}()}
}
\concept{postgresql functions}
