% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{sbfx_load_datas_from_pg}
\alias{sbfx_load_datas_from_pg}
\title{Load data frames from a PostgreSQL database}
\usage{
sbfx_load_datas_from_pg(
  schema = getOption("psql.schema", "public"),
  rename = identity,
  env = parent.frame(),
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.config_value", "default")
)
}
\arguments{
\item{schema}{A string of the schema name. Default value is \code{"public"}.}

\item{rename}{A single function argument which takes a character vector
and returns a character vector of the same length.
Used to rename objects before they are loaded into the environment.}

\item{env}{The environment to  the objects into}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option
and uses \code{NULL} if no value supplied.}

\item{config_value}{A string of the name of value. The default value grabs
the value from the psql.config_value option and uses \code{"default"} if no
value is supplied.}
}
\value{
An invisible character vector of the paths to the saved objects.
}
\description{
Load all the tables in a PostgreSQL database schema into R as data frames.
}
\examples{
\dontrun{
sbfx_load_datas_from_pg()
sbfx_load_datas_from_pg(schema = "capture")
sbfx_load_datas_from_pg(rename = toupper)
}
}
\seealso{
Other postgresql functions: 
\code{\link{sbfx_backup_pg}()},
\code{\link{sbfx_close_pg}()},
\code{\link{sbfx_create_pg}()},
\code{\link{sbfx_execute_pg}()},
\code{\link{sbfx_get_config_file}()},
\code{\link{sbfx_get_config_value}()},
\code{\link{sbfx_get_schema}()},
\code{\link{sbfx_list_tables_pg}()},
\code{\link{sbfx_load_data_from_pg}()},
\code{\link{sbfx_open_pg}()},
\code{\link{sbfx_reset_config_file}()},
\code{\link{sbfx_reset_config_value}()},
\code{\link{sbfx_reset_schema}()},
\code{\link{sbfx_save_data_to_pg}()},
\code{\link{sbfx_set_config_file}()},
\code{\link{sbfx_set_config_value}()},
\code{\link{sbfx_set_schema}()}

Other load functions: 
\code{\link{sbfx_load_data_from_pg}()}
}
\concept{load functions}
\concept{postgresql functions}
