% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{sbfx_open_pg}
\alias{sbfx_open_pg}
\title{Open PostgreSQL connection}
\usage{
sbfx_open_pg(
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.config_value", "default")
)
}
\arguments{
\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option
and uses \code{NULL} if no value supplied.}

\item{config_value}{A string of the name of value. The default value grabs
the value from the psql.config_value option and uses \code{"default"} if no
value is supplied.}
}
\value{
An S4 object that inherits from DBIConnection.
}
\description{
Connect to a PostgreSQL database with a config.yml file.
}
\details{
Wrapper on \code{psql::psql_connect()}.
}
\examples{
\dontrun{
conn <- sbfx_open_pg()
sbfx_close_pg(conn)

sbfx_open_pg("config.yml")
sbfx_close_pg(conn)

sbfx_open_pg(config_path = "config.yml", config_value = "database")
sbfx_close_pg(conn)
}
}
\seealso{
Other postgresql functions: 
\code{\link{sbfx_backup_pg}()},
\code{\link{sbfx_close_pg}()},
\code{\link{sbfx_create_pg}()},
\code{\link{sbfx_execute_pg}()},
\code{\link{sbfx_get_config_file}()},
\code{\link{sbfx_get_config_value}()},
\code{\link{sbfx_get_schema}()},
\code{\link{sbfx_list_tables_pg}()},
\code{\link{sbfx_load_data_from_pg}()},
\code{\link{sbfx_load_datas_from_pg}()},
\code{\link{sbfx_reset_config_file}()},
\code{\link{sbfx_reset_config_value}()},
\code{\link{sbfx_reset_schema}()},
\code{\link{sbfx_save_data_to_pg}()},
\code{\link{sbfx_set_config_file}()},
\code{\link{sbfx_set_config_value}()},
\code{\link{sbfx_set_schema}()}
}
\concept{postgresql functions}
