% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{sbfx_reset_schema}
\alias{sbfx_reset_schema}
\title{Reset schema name}
\usage{
sbfx_reset_schema()
}
\value{
An invisible string of the schema name the database is set to.
}
\description{
Reset schema name back to public
}
\examples{
\dontrun{
sbfx_reset_schema()
}
}
\seealso{
Other postgresql functions: 
\code{\link{sbfx_backup_pg}()},
\code{\link{sbfx_close_pg}()},
\code{\link{sbfx_create_pg}()},
\code{\link{sbfx_execute_pg}()},
\code{\link{sbfx_get_config_file}()},
\code{\link{sbfx_get_config_value}()},
\code{\link{sbfx_get_schema}()},
\code{\link{sbfx_list_tables_pg}()},
\code{\link{sbfx_load_data_from_pg}()},
\code{\link{sbfx_load_datas_from_pg}()},
\code{\link{sbfx_open_pg}()},
\code{\link{sbfx_reset_config_file}()},
\code{\link{sbfx_reset_config_value}()},
\code{\link{sbfx_save_data_to_pg}()},
\code{\link{sbfx_set_config_file}()},
\code{\link{sbfx_set_config_value}()},
\code{\link{sbfx_set_schema}()}
}
\concept{postgresql functions}
