% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{sbfx_save_data_to_pg}
\alias{sbfx_save_data_to_pg}
\title{Add data frame to PostgreSQL database}
\usage{
sbfx_save_data_to_pg(
  x,
  x_name = NULL,
  schema = getOption("psql.schema", "public"),
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.config_value", "default")
)
}
\arguments{
\item{x}{The data frame to save.}

\item{x_name}{A string of the name.}

\item{schema}{A string of the schema name. Default value is \code{"public"}.}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option
and uses \code{NULL} if no value supplied.}

\item{config_value}{A string of the name of value. The default value grabs
the value from the psql.config_value option and uses \code{"default"} if no
value is supplied.}
}
\value{
A scalar numeric.
}
\description{
Add data in a data frame to your PostgreSQL database. The data frame name
must match the table name in your database, if not use the \code{tbl_name}
argument to pass the table name.
}
\details{
Wrapper on \code{psql::psql_add_data()}.
}
\examples{
\dontrun{
sbfx_save_data_to_pg(outing, "creel")
sbfx_save_data_to_pg(outing_new, "creel", "outing")
}
}
\seealso{
Other postgresql functions: 
\code{\link{sbfx_backup_pg}()},
\code{\link{sbfx_close_pg}()},
\code{\link{sbfx_create_pg}()},
\code{\link{sbfx_execute_pg}()},
\code{\link{sbfx_get_config_file}()},
\code{\link{sbfx_get_config_value}()},
\code{\link{sbfx_get_schema}()},
\code{\link{sbfx_list_tables_pg}()},
\code{\link{sbfx_load_data_from_pg}()},
\code{\link{sbfx_load_datas_from_pg}()},
\code{\link{sbfx_open_pg}()},
\code{\link{sbfx_reset_config_file}()},
\code{\link{sbfx_reset_config_value}()},
\code{\link{sbfx_reset_schema}()},
\code{\link{sbfx_set_config_file}()},
\code{\link{sbfx_set_config_value}()},
\code{\link{sbfx_set_schema}()}
}
\concept{postgresql functions}
\concept{save functions}
