% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{sbfx_set_config_value}
\alias{sbfx_set_config_value}
\title{Set config value}
\usage{
sbfx_set_config_value(value = NULL)
}
\arguments{
\item{value}{A string of the config file value to grab.}
}
\value{
An invisible string of the value given.
}
\description{
Wrapper for setting the \code{psql.config_value} options parameter.
}
\details{
This function is recommended to be added to your header when used.
}
\examples{
\dontrun{
sbfx_set_config_value("shinyapp")
}
}
\seealso{
Other postgresql functions: 
\code{\link{sbfx_backup_pg}()},
\code{\link{sbfx_close_pg}()},
\code{\link{sbfx_create_pg}()},
\code{\link{sbfx_execute_pg}()},
\code{\link{sbfx_get_config_file}()},
\code{\link{sbfx_get_config_value}()},
\code{\link{sbfx_get_schema}()},
\code{\link{sbfx_list_tables_pg}()},
\code{\link{sbfx_load_data_from_pg}()},
\code{\link{sbfx_load_datas_from_pg}()},
\code{\link{sbfx_open_pg}()},
\code{\link{sbfx_reset_config_file}()},
\code{\link{sbfx_reset_config_value}()},
\code{\link{sbfx_reset_schema}()},
\code{\link{sbfx_save_data_to_pg}()},
\code{\link{sbfx_set_config_file}()},
\code{\link{sbfx_set_schema}()}
}
\concept{postgresql functions}
