% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfoldr2ext-package.R
\docType{package}
\name{subfoldr2ext-package}
\alias{subfoldr2ext}
\alias{subfoldr2ext-package}
\title{subfoldr2ext: Extra Functions Related to subfoldr2}
\description{
Extra or externally based functions that are related to functions in subfoldr2.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/subfoldr2ext}
  \item Report bugs at \url{https://github.com/poissonconsulting/subfoldr2ext/issues}
}

}
\author{
\strong{Maintainer}: Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})

Authors:
\itemize{
  \item Duncan Kennedy \email{duncan@poissonconsulting.ca} (\href{https://orcid.org/0009-0001-4880-5751}{ORCID})
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
}

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder]
}

}
\keyword{internal}
